+++ 8680s_Lua_Pack/CHANGELOG.md
@@ -4,7 +4,8 @@
 * Made match_tag able to filter on the tag’s contents being a superset of
   another tag.
 * Added unmodify library.
-* Added wml/{is_cfg,is_subtag,is_child,is_unit_proxy,to_unit_cfg} functions.
+* Added to wml library: is_tag, is_cfg, to_tag, to_cfg, is_subtag, is_child,
+  is_unit_proxy, to_unit_cfg.
 * Added utils/{AND,OR} constants.
 
 0.7.0