+++ 8680s_Lua_Pack/CHANGELOG.md
@@ -1,11 +1,17 @@
 
+0.8.0
+===============================================================================
+* Added modifications library, with:
+** functions remove_effect, remove_object;
+** tag [remove_object].
+
 0.7.1
 ===============================================================================
 * Made match_tag able to filter on the tag’s contents being a superset of
   another tag.
-* Added unmodify library.
-* Added to wml library: is_tag, is_cfg, to_tag, to_cfg, is_subtag, is_child,
-  is_unit_proxy, to_unit_cfg.
+* Added to wml library:
+** functions is_tag, is_cfg, to_tag, to_cfg, is_subtag, is_child,
+   is_unit_proxy, to_unit_cfg.
 * Added utils/{AND,OR} constants.
 
 0.7.0